High Volume Data Loader Use Case
A walkthrough on processing very large CSV uploads and database inserts in a SpringBoot application. It will also discuss the implementation of a custom multipart resolver to limit the number of file uploads, along with the incorporation of a Rate Limiter. Additionally, methods for measuring performance characteristics using Java Mission Control, SQL Server Extended Events, and Apache JMeter will be explored.